Add files via upload
The goal is to analyze the Google Play Store Apps dataset to derive insights into
the app market. We aim to understand the factors that contribute to an app's
success, including its user ratings, reviews, and category. We want to explore user
sentiments towards apps by analyzing the user reviews dataset. Additionally, we
can see insights into the popularity of app categories based on the total number
of installs and the sentiment polarity of user reviews. The ultimate objective is to
provide recommendations for app developers to enhance their app's
performance and user satisfaction.